# Improve supplier list query performance